CMake files are added and Build section is added to README.md
Three CMakeLists.txt files are added in lumen, lumen/src and
lumen/Dependencies directories, those files will build the glm and SFML
librarise and link with the main executable file.
The main.cpp in lumen/src is only a test example to show the SFML and
glm are built and linked correctly.
A "Build" section is added to README.md file in lumen.